+++ receiver/kafkareceiver/factory_test.go
@@ -80,6 +80,50 @@ func TestWithTracesUnmarshalers(t *testing.T) {
 	})
 }
 
+func TestCreateMetricsReceiver(t *testing.T) {
+	cfg := createDefaultConfig().(*Config)
+	cfg.Brokers = []string{"invalid:9092"}
+	cfg.ProtocolVersion = "2.0.0"
+	f := kafkaReceiverFactory{metricsUnmarshalers: defaultMetricsUnmarshalers()}
+	r, err := f.createMetricsReceiver(context.Background(), componenttest.NewNopReceiverCreateSettings(), cfg, nil)
+	// no available broker
+	require.Error(t, err)
+	assert.Nil(t, r)
+}
+
+func TestCreateMetricsReceiver_error(t *testing.T) {
+	cfg := createDefaultConfig().(*Config)
+	cfg.ProtocolVersion = "2.0.0"
+	// disable contacting broker at startup
+	cfg.Metadata.Full = false
+	f := kafkaReceiverFactory{metricsUnmarshalers: defaultMetricsUnmarshalers()}
+	r, err := f.createMetricsReceiver(context.Background(), componenttest.NewNopReceiverCreateSettings(), cfg, nil)
+	require.NoError(t, err)
+	assert.NotNil(t, r)
+}
+
+func TestWithMetricsUnmarshalers(t *testing.T) {
+	unmarshaler := &customMetricsUnmarshaler{}
+	f := NewFactory(WithMetricsUnmarshalers(unmarshaler))
+	cfg := createDefaultConfig().(*Config)
+	// disable contacting broker
+	cfg.Metadata.Full = false
+	cfg.ProtocolVersion = "2.0.0"
+
+	t.Run("custom_encoding", func(t *testing.T) {
+		cfg.Encoding = unmarshaler.Encoding()
+		receiver, err := f.CreateMetricsReceiver(context.Background(), componenttest.NewNopReceiverCreateSettings(), cfg, nil)
+		require.NoError(t, err)
+		require.NotNil(t, receiver)
+	})
+	t.Run("default_encoding", func(t *testing.T) {
+		cfg.Encoding = defaultEncoding
+		receiver, err := f.CreateMetricsReceiver(context.Background(), componenttest.NewNopReceiverCreateSettings(), cfg, nil)
+		require.NoError(t, err)
+		assert.NotNil(t, receiver)
+	})
+}
+
 func TestCreateLogsReceiver(t *testing.T) {
 	cfg := createDefaultConfig().(*Config)
 	cfg.Brokers = []string{"invalid:9092"}
@@ -127,6 +171,9 @@ func TestWithLogsUnmarshalers(t *testing.T) {
 type customTracesUnmarshaler struct {
 }
 
+type customMetricsUnmarshaler struct {
+}
+
 type customLogsUnmarshaler struct {
 }
 
@@ -137,6 +184,14 @@ func (c customTracesUnmarshaler) Unmarshal([]byte) (pdata.Traces, error) {
 }
 
 func (c customTracesUnmarshaler) Encoding() string {
+	return "custom"
+}
+
+func (c customMetricsUnmarshaler) Unmarshal([]byte) (pdata.Metrics, error) {
+	panic("implement me")
+}
+
+func (c customMetricsUnmarshaler) Encoding() string {
 	return "custom"
 }